requirements.txt
amqp==2.6.0
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-SocketIO==4.3.1
Flask-WTF==0.14.3
flower==0.9.5
future==0.18.2
gevent==20.6.2
gevent-websocket==0.10.1
greenlet==0.4.16
humanize==2.5.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
Pillow==7.2.0
prometheus-client==0.8.0
python-chess==0.31.3
python-engineio==3.13.1
python-socketio==4.6.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
rom==1.0.0
six==1.15.0
tornado==6.0.4
urllib3==1.25.10
vine==1.3.0
Werkzeug==1.0.1
WTForms==2.3.2
zope.event==4.4
zope.interface==5.1.0
